Move some functions from util and math to new array module
