+++ scico/operator/biconvolve.py
@@ -13,10 +13,10 @@ import numpy as np
 from jax.scipy.signal import convolve
 
 from scico._generic_operators import LinearOperator, Operator
+from scico.array import is_nested
 from scico.blockarray import BlockArray
 from scico.linop import Convolve, ConvolveByX
 from scico.typing import BlockShape, DType, JaxArray
-from scico.util import is_nested
 
 __author__ = """Luke Pfister <luke.pfister@gmail.com>"""
